freeze_requirements.txt
absl-py==2.1.0
accelerate==0.34.2
annotated-types==0.7.0
asttokens==2.4.1
av==13.1.0
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
colorlog==6.8.2
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
decord==0.6.0
deepspeed==0.15.1
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
filelock @ file:///croot/filelock_1700591183607/work
flash-attn==2.6.3
fsspec==2024.9.0
ftfy==6.3.1
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
grpcio==1.66.2
hjson==3.1.0
huggingface-hub==0.25.2
idna @ file:///croot/idna_1714398848350/work
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1716993405101/work
jupyter_client==8.6.3
jupyter_core==5.7.2
Levenshtein==0.26.0
Markdown==3.7
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib-inline==0.1.7
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
mpmath @ file:///croot/mpmath_1690848262763/work
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1717597493534/work
ninja==1.11.1.1
numpy==1.26.4
open_clip_torch==2.29.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
peft==0.13.1
pexpect==4.9.0
pillow @ file:///croot/pillow_1721059439630/work
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.0
pytz==2024.2
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq==26.2.0
qwen-vl-utils==0.0.8
RapidFuzz==3.10.0
regex==2024.9.11
requests @ file:///croot/requests_1721410876868/work
safetensors==0.4.5
sentencepiece==0.2.0
six==1.16.0
stack-data==0.6.3
sympy @ file:///croot/sympy_1724938189289/work
tensorboard==2.18.0
tensorboard-data-server==0.7.2
timm==1.0.9
tokenizers==0.19.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers@21fac7abba2a37fae86106f87fcf9974fd1e3830
triton==2.1.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.2
urllib3 @ file:///croot/urllib3_1727769808118/work
wcwidth==0.2.13
Werkzeug==3.0.4
